Call correct function pass release arg

+++ velocity/compile_velocity_cpu.py
@@ -13,13 +13,14 @@ from utils import (
     tmp_call_13_to_reduction,
     levmask_to_reduction,
     compare_got_and_want,
-    compile,
+    compile_sdfg,
     count_loops,
 )
 
 
 use_cache = True
 run_benchmark = False
+cleanup = False
 
 # Load SDFG
 sdfg = dace.SDFG.from_file("velocity.sdfgz")
@@ -87,7 +88,7 @@ for node, state in sdfg.all_nodes_recursive():
 ################################################################################
 
 # Compile the SDFG
-compile(sdfg, gpu=False)
+compile_sdfg(sdfg, gpu=False, release=False)
 
 # check if execution was successful
 if os.system(f"./{sdfg_name}") != 0:
@@ -124,12 +125,13 @@ if run_benchmark:
 ################################################################################
 
 # remove the compiled program
-os.remove(sdfg_name)
+if cleanup:
+    os.remove(sdfg_name)
 
-# remove .got and .want files
-for f in os.listdir():
-    if f.endswith(".got") or f.endswith(".want"):
-        os.remove(f)
+    # remove .got and .want files
+    for f in os.listdir():
+        if f.endswith(".got") or f.endswith(".want"):
+            os.remove(f)
 
-# remove the .dacecache folder
-shutil.rmtree(build_loc)
+    # remove the .dacecache folder
+    shutil.rmtree(build_loc)
